chore: refactor AlertBanner to individual files to show docs
