+++ packages/kit/src/alert-banner/index.ts
@@ -1,1 +1,4 @@
 export * from "./AlertBanner";
+export { AlertBannerRoot } from "./AlertBannerRoot";
+export { AlertBannerTrigger } from "./AlertBannerTrigger";
+export { AlertBannerContent } from "./AlertBannerContent";
